% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/neuralnet.R
\name{calculateVals}
\alias{calculateVals}
\title{Calculate the Values of Vertices in Neural Network Model}
\usage{
calculateVals(g, v.index)
}
\arguments{
\item{g,}{the graph model}

\item{v.index,}{the index of a vertex}
}
\value{
an updated graph model
}
\description{
For a given vertex, input and the output signals are calculated.
}

