% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/propagation.R
\name{checkAttributes}
\alias{checkAttributes}
\title{Check if updated attribute is present}
\usage{
checkAttributes(g, S)
}
\arguments{
\item{g}{igraph graph object}
}
\description{
Check if updated attribute is present
}

