% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc-utilities.R
\name{checkVertex}
\alias{checkVertex}
\title{Type checking for igraph vertex arguments.}
\usage{
checkVertex(v)
}
\arguments{
\item{v}{Assumed to represent a vertex.

This function, returns an error if v is not of the class 'character'.}
}
\description{
igraph fnctions that use igraph vertices as arguments can typicall take
the vertex object itself (igraph.vs), the vertex id (numeric),
or the vertex name (character).  This flexibility can lead to unsafe code.
The convention in this package is to use the vertex name.
}

