% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc-utilities.R
\name{detect_cycle}
\alias{detect_cycle}
\title{Find a cycle
Returns the first cycle detected in the graph.  Relies on shortest path calculations.
If there is a "weight" edge attribute, it will be incorporated in the shortest edge
calculation.}
\usage{
detect_cycle(g)
}
\description{
Find a cycle
Returns the first cycle detected in the graph.  Relies on shortest path calculations.
If there is a "weight" edge attribute, it will be incorporated in the shortest edge
calculation.
}

