% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/neuralnet.R
\name{doChainRule}
\alias{doChainRule}
\title{Calculates the Derivative of a Node's Output Signal w.r.t a Weight.}
\usage{
doChainRule(g, v, e)
}
\arguments{
\item{g}{a model}

\item{v}{vertex index}

\item{e}{edge index}
}
\value{
a vector corresponding to the Derivative
}
\description{
Calculates the Derivative of a Node's Output Signal w.r.t a Weight.
}

