% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/propagation.R
\name{edge_updater}
\alias{edge_updater}
\title{Edge propagation instance
Executes a callback on the edge that calculates a value.  The callback can rely
on values of other edges that are calculated by the same callback.  When the function
is applied to an edge, it checks that an 'update' attribute on all the edges that it 
relies on is set to true before applying the callback.  If not, the callback is applied
to those unupdated edges first.}
\usage{
edge_updater(g, e, get_determiners, callback, verbose = FALSE)
}
\arguments{
\item{g}{an igraph object}

\item{e}{an index for an edge in g}

\item{get_determiners}{a function that returns the indices of the edges that must be updated
before the callback is executed.}

\item{verbose}{if TRUE prints out details of propagation}

\item{a}{function that performs an operation on an edge.}
}
\description{
Edge propagation instance
Executes a callback on the edge that calculates a value.  The callback can rely
on values of other edges that are calculated by the same callback.  When the function
is applied to an edge, it checks that an 'update' attribute on all the edges that it 
relies on is set to true before applying the callback.  If not, the callback is applied
to those unupdated edges first.
}

