% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/neuralnet.R
\name{examineFFGraph}
\alias{examineFFGraph}
\title{Summarize the values of an igraph neural net model}
\usage{
examineFFGraph(g, formatVertexAttr = formatVertexList)
}
\description{
Summarize the values of an igraph neural net model
}

