% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc-utilities.R
\name{first_v}
\alias{first_v}
\alias{last_v}
\alias{random_v}
\title{Select vertices from a sequence}
\usage{
first_v(g, s)

last_v(g, s)

random_v(g, s)
}
\arguments{
\item{g}{igraph object}

\item{s}{an object of class igraph.vs}
}
\value{
the index of a vertex from the sequence
}
\description{
Given a vertex sequence (class igraph.vs), selects the first, last, or a random vertex from
the sequence
}

