% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph-sim.R
\name{fit_barabasi_power}
\alias{fit_barabasi_power}
\title{Use input graph to estimate power of preferential attachment for Barabasi-Albert algorithm.
Fits a power law to the input graph and returns the value to use as the power of
preferential attachment in a Barabasi-Albert simulation of a scale-free network.  This
is the 'power' argument in igraph's 'barabasi.game" function.}
\usage{
fit_barabasi_power(g, mle = FALSE)
}
\arguments{
\item{g}{an input igraph object, the power law will be simulated from this object.}

\item{mle}{if TRUE, estimate the power value of the out-degree distribution using mle via
BFGS optimization.  If FALSE, use the 'plfit' implementation (see the power.law.fit function
in igraph).  Default is FALSE.  FALSE option will generate a warning if a bad fit is detected
but the TRUE option will not.}
}
\value{
a numeric value.
}
\description{
Use input graph to estimate power of preferential attachment for Barabasi-Albert algorithm.
Fits a power law to the input graph and returns the value to use as the power of
preferential attachment in a Barabasi-Albert simulation of a scale-free network.  This
is the 'power' argument in igraph's 'barabasi.game" function.
}

