% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/graph-sim.R
\name{generateMultiConnectedDAG}
\alias{generateMultiConnectedDAG}
\title{Generate Multi-connected DAG}
\usage{
generateMultiConnectedDAG(n)
}
\arguments{
\item{n}{Number of nodes}
}
\value{
An igraph object
}
\description{
An alternative to the scale-free DAG generating Barabasi algorithm in igraph..
This uses Ide's and Cozman's DAG algorithm in the bnlearn package, which generates DAGS with
more connectivity.  This function currently only outputs one leaf node.
}
\examples{
g <- generateMultiConnectedDag(10)
igraphVizPlot(g)
}

