% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lucy.R
\name{isBUpstreamOfA}
\alias{get_connecting_edges}
\alias{get_connecting_nodes}
\alias{isBUpstreamOfA}
\title{Find All Vertices or Edges on Paths Connecting Two Vertices}
\usage{
isBUpstreamOfA(g, a, b)

get_connecting_nodes(g, src, trg)

get_connecting_edges(g, src, trg)
}
\arguments{
\item{g}{igraph object}

\item{src}{vertex index of class numeric or igraph.vs, starting vertex}

\item{trg}{vertex index of class numeric or igraph.vs, ending vertex}
}
\value{
numeric vertex indices, including the src and trg indices
}
\description{
Given two vertices, find all vertices lying on all paths between those two 
vertices.
}

