% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lucy.R
\name{get_downstream_nodes}
\alias{get_downstream_nodes}
\alias{get_upstream_nodes}
\title{Find all Vertices Upstream or Downstream of a Given Vertex in Directed Graph}
\usage{
get_downstream_nodes(g, w)

get_upstream_nodes(g, w)
}
\arguments{
\item{g}{igraph object}

\item{w}{character, the name of a vertex in g}
}
\value{
character vector of vertex names.
}
\description{
Find all Vertices Upstream or Downstream of a Given Vertex in Directed Graph
}

