% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc-utilities.R
\name{get_roots}
\alias{get_leaves}
\alias{get_roots}
\title{Find all the leaves and root vertices in a directed graph.
A root node has no parents}
\usage{
get_roots(g)

get_leaves(g)
}
\arguments{
\item{g}{an igraph directed graph}
}
\value{
chacter vector of vertex names
}
\description{
Find all the leaves and root vertices in a directed graph.
A root node has no parents
}
\examples{
g <- ba.game(10)
get_roots(g)
}

