% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/igraphr.R
\name{isBUpstreamOfA}
\alias{getConnectingEdges}
\alias{getConnectingNodes}
\alias{isBUpstreamOfA}
\title{Find All Vertices or Edges on Paths Connecting Two Vertices}
\usage{
isBUpstreamOfA(g, a, b)

getConnectingNodes(g, src, trg)

getConnectingEdges(g, src, trg)
}
\arguments{
\item{g}{igraph object}

\item{src}{vertex index of class numeric or igraph.vs, starting vertex}

\item{trg}{vertex index of class numeric or igraph.vs, ending vertex}
}
\value{
numeric vertex indices, including the src and trg indices
}
\description{
Given two vertices, find all vertices lying on all paths between those two
vertices.
}

