% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/lucy.R
\name{getDownstreamNodes}
\alias{getDownstreamNodes}
\alias{getUpstreamNodes}
\title{Find all Vertices Upstream or Downstream of a Given Vertex in Directed Graph}
\usage{
getDownstreamNodes(g, w)

getUpstreamNodes(g, w)
}
\arguments{
\item{g}{igraph object}

\item{w}{character, the name of a vertex in g}
}
\value{
character vector of vertex names.
}
\description{
Find all Vertices Upstream or Downstream of a Given Vertex in Directed Graph
}

