% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/misc-utilities.R
\name{getEdgeVertex}
\alias{getEdgeVertex}
\title{Pull the Vertex from a Given Edge in Graph
Returns the index of a vertex in an edge}
\usage{
getEdgeVertex(g, e, node = c("from", "to"))
}
\arguments{
\item{g}{igraph object}

\item{e}{integer/numeric/igraph.es edge index}

\item{node}{'from' for from node, 'to' for to-node}
}
\value{
numeric of the vertex index
}
\description{
Pull the Vertex from a Given Edge in Graph
Returns the index of a vertex in an edge
}

