% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/misc-utilities.R
\name{getLeaves}
\alias{getLeaves}
\title{Find all the leaf vertices in a directed graph.
A root node has no children}
\usage{
getLeaves(g)
}
\arguments{
\item{g}{an igraph directed graph}
}
\value{
chacter vector of vertex names
}
\description{
Find all the leaf vertices in a directed graph.
A root node has no children
}
\examples{
g <- ba.game(10)
getLeaves(g)
}

