% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/neuralnet.R
\name{getLinearCombination}
\alias{getLinearCombination}
\title{A Simple Matrix Multiplication to Calculate Linear Inputs}
\usage{
getLinearCombination(weights, model.mat)
}
\description{
A Simple Matrix Multiplication to Calculate Linear Inputs
}

