% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/propagation.R
\name{getOrdering}
\alias{getOrdering}
\title{Graph Traversal}
\usage{
getOrdering(g, S, traversal = "depth.first")
}
\description{
This uses the igraph traversal algorithms to produce a perhaps a better ordering of
vertexs in the traversal done in the updating.  Not yet used, may be applied later to
improve speed, or when a callback is needed to stop the traversal in the event of cycles.
}

