% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/misc-utilities.R
\name{getRoots}
\alias{getLeaves}
\alias{getRoots}
\title{Find all the leaves and root vertices in a directed graph.
A root node has no parents}
\usage{
getRoots(g)

getLeaves(g)
}
\arguments{
\item{g}{an igraph directed graph}
}
\value{
chacter vector of vertex names
}
\description{
Find all the leaves and root vertices in a directed graph.
A root node has no parents
}
\examples{
g <- ba.game(10)
getRoots(g)
}

