% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/propagation.R
\name{getTraverser}
\alias{getTraverser}
\title{Traversal for Propogation}
\usage{
getTraverser(iterator)
}
\arguments{
\item{iterator}{either "edge", or "vertex"
E if traversing over edges, or V if traversing over vertices}
}
\description{
This closure creates a function that traverses the graph, applying the updater as it travels.
}

