% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/propagation.R
\name{getUpdater}
\alias{getUpdater}
\title{Propagator closure}
\usage{
getUpdater(iterator)
}
\arguments{
\item{iterator}{either "edge", or "vertex"}
}
\value{
A function that updates edges (S = E) or vertices (S = V)
}
\description{
A closure generates functions for doing propagations on a graph.
}

