% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph-plots.R
\name{igraphviz}
\alias{igraphviz}
\title{igraph plotting with Graphviz library}
\usage{
igraphviz(g, main = NULL)
}
\arguments{
\item{g}{an igraph object.}

\item{main}{character, the desired title of the graph. Default is NULL}
}
\value{
graphviz.plot returns invisibly the graph object produced by Rgraphviz.
}
\description{
A simple plotter of igrpah objects with RGraphviz.  Nodes are labeled by 
their vertex names.  For more complicated graphing is need, use Graphviz 
directly.
}
\examples{
g <- ba.game(20)
igraphviz(g)
g <- layer_DAGs(3, 4)
igraphviz(g)
}

