% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc-utilities.R
\name{in_degree}
\alias{in_degree}
\alias{out_degree}
\title{In/Out-degree of a set of vertices in a directed graph}
\usage{
in_degree(g, v.set)

out_degree(g, v.set)
}
\arguments{
\item{g}{an igraph directed graph}

\item{v.set}{a set of vertex names}
}
\value{
numeric vector of degree counts
}
\description{
In/Out-degree of a set of vertices in a directed graph
}
\examples{
g <- ba.game(10)
inDegree(g, V(g))
}

