% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/neuralnet.R
\name{initializeGraph}
\alias{initializeGraph}
\title{Primes a Graph Object for Fitting a Neural Network Model}
\usage{
initializeGraph(g, input.table, output.table, activation = logistic,
  activation.prime = logistic.prime, min.max.constraints = NULL)
}
\arguments{
\item{g}{igraph object with vertices corresponding to input nodes, output nodes, and
hidden nodes.}

\item{input.table}{data frame of the input values}

\item{output.table}{data frame of the output values}

\item{activation}{function, the desired activation function}

\item{activation.prime,}{the Derivative of the desired activation function}

\item{min.max.constraints}{(optional) numeric containing the limiting range of the estimates}
}
\value{
A graph with all the attributes needed to fit the neural network model.
}
\description{
Primes a Graph Object for Fitting a Neural Network Model
}

