% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/neuralnet.R
\name{initializeVertexVectors}
\alias{initializeVertexVectors}
\title{Initialize Numerically-Valued Vertex Attributes
For a given vertex, vertex attibutes that take a vector as a value are
initialized with a placeholder.}
\usage{
initializeVertexVectors(g, v.index)
}
\arguments{
\item{g}{graph model}

\item{v.index}{vertex index}
}
\value{
graph object with placeholders for numerically-valued vertex attributes
}
\description{
Initialize Numerically-Valued Vertex Attributes
For a given vertex, vertex attibutes that take a vector as a value are
initialized with a placeholder.
}

