% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/misc-utilities.R
\name{iparents}
\alias{ichildren}
\alias{iparents}
\title{#Extends Magrittr's "$" Pipe Operator igraph.
#
#For more details Magrittr see \code{?magrittr}
#
#}
\usage{
iparents(g, v)

ichildren(g, v)
}
\arguments{
\item{g}{igraph graph object, directed}

\item{v}{numeric vertex index or igraph.vs object}
}
\description{
#Extends Magrittr's "$" Pipe Operator igraph.
#
#For more details Magrittr see \code{?magrittr}
#
#
}
\examples{
#library(igraphr)
#g <- erdos.renyi.game(100, 2/100)
#V(g)$color <- "green"
#g \%>\% V \%$\% color
#
}

