% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lucy.R
\name{isBDownstreamOfA}
\alias{isBDownstreamOfA}
\title{Test if a Vertex is Downstream or Upstream of Another}
\usage{
isBDownstreamOfA(g, a, b)
}
\arguments{
\item{g}{igraph object}

\item{a}{numeric index or igraph.vs object, source vertex}

\item{end}{vertex}
}
\value{
TRUE/FALSE value
}
\description{
Test if a Vertex is Downstream or Upstream of Another
}

