% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/neuralnet.R
\name{logistic}
\alias{logistic}
\alias{logistic.prime}
\title{The Logistic Function and Its Derivative}
\usage{
logistic(z)

logistic.prime(z)
}
\description{
The Logistic Function and Its Derivative
}

