% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc-utilities.R
\name{name_vertices}
\alias{name_edges}
\alias{name_vertices}
\title{Quick Naming of the Vertices/Edges in a Graph}
\usage{
name_vertices(g)

name_edges(g)
}
\arguments{
\item{g}{an igraph directed graph}
}
\value{
the graph with the vertex attribute 'name'
}
\description{
If the vertex/edges of an igraph object have no name, quickly add a name 
based on the vertex index/vertices in the edge. For vertices the name is the 
vertex id, the edges it is 'from->to'.
}

