% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/misc-utilities.R
\name{outDegree}
\alias{outDegree}
\title{Out degree of a set of vertices in a directed graph}
\usage{
outDegree(g, v.set)
}
\arguments{
\item{g}{an igraph directed graph}

\item{v.set}{a set of vertex names}
}
\value{
numeric vector of degree counts
}
\description{
Out degree of a set of vertices in a directed graph
}
\examples{
g <- ba.game(10)
inDegree(g, V(g))
}

