% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph-sim.R
\name{power_law_sim}
\alias{power_law_sim}
\title{Simulate a scale-free network given an input network.}
\usage{
power_law_sim(g, n)
}
\arguments{
\item{n}{number of desired vertices in the output graph}

\item{input}{graph}
}
\value{
a new igraph object.
}
\description{
Fit a power law to the input network, and simulate a new network based on 
that fitted power law.  Assumes the input network has a power law degree distribution.
}

