% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/neuralnet.R
\name{resetUpdateAttributes}
\alias{resetUpdateAttributes}
\title{Reset Model Attributes}
\usage{
resetUpdateAttributes(g)
}
\arguments{
\item{g}{a model}
}
\value{
A model with updated attributes reset to FALSE
}
\description{
Resets the attributes of a model, after some of the attributes have been updated.
}

