% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc-utilities.R
\name{reverse_edge}
\alias{reverse_edge}
\title{Reverse an edge in a directed graph}
\usage{
reverse_edge(g, e)
}
\arguments{
\item{g}{a directed igraph object}

\item{e}{an edge (integer index or "igraph.es" object)}
}
\value{
a new igraph object, the old graph with edges reverse.
}
\description{
Return a new directed graph instance a given edge oriented in the opposite direction 
relative to the input graph.  Will not preserve graph, vertex,
and edge attributes.
}

