% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc-utilities.R
\name{reverse_edges}
\alias{reverse_edges}
\title{Reverse the edges of a directed graph}
\usage{
reverse_edges(g)
}
\arguments{
\item{g}{a directed igraph object}
}
\value{
a new igraph object, the old graph with edges reverse.
}
\description{
Return a new directed graph instance with each edge oriented in the opposite direction 
relative to the corresponding edge in the input graph.  Will not preserve graph, vertex,
and edge attributes.
}

