% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc-utilities.R
\name{setIterator}
\alias{setIterator}
\title{Create an igraph Iterator Substitution Function}
\usage{
setIterator(iterator)
}
\arguments{
\item{iterator}{either "edge", or "vertex"}
}
\description{
Takes either igraph iterator functions \code{E} and \code{V}, and creates a mirror function \code{S}
that can be used in code intended to work on either edges or vertices. Also creates the replacement 
function.  Creates the object in the parent environment.
}

