% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/misc-utilities.R
\name{with.igraph.vs}
\alias{with.igraph.vs}
\title{Extends Magrittr's "$" Pipe Operator igraph.}
\usage{
\method{with}{igraph.vs}(data, expr, ...)
}
\description{
For more details Magrittr see \code{?magrittr}
}
\examples{
library(igraphr)
g <- erdos.renyi.game(100, 2/100)
V(g)$color <- "green"
g \%>\% V \%$\% color
}

