% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc-utilities.R
\name{iparents}
\alias{ichildren}
\alias{imb}
\alias{iparents}
\title{Parents/Children/Markov Blanket of a Vertex}
\usage{
iparents(g, v)

ichildren(g, v)

imb(g, v)
}
\arguments{
\item{g}{igraph graph object, directed}

\item{v}{numeric vertex index or igraph.vs object}
}
\description{
Parents/Children/Markov Blanket of a Vertex
}

