% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/misc-utilities.R
\name{nameVertices}
\alias{nameVertices}
\title{Quick Naming of the Vertices/Edges in a Graph}
\usage{
nameVertices(g, v.set = V(g))
}
\arguments{
\item{g}{an igraph directed graph}

\item{v.set}{indices for the vertices that are to be named}

\item{e.set}{indices for the edges that that are to be named}
}
\value{
chacter vector of vertex names
}
\description{
Quickly add a name attribute to the vertices or edges in the graph.
For vertices the name is the vertex id, the edges it is 'from->to'.
}

