% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph-plots.R
\name{plot_path}
\alias{plot_path}
\title{Plot path from an upstream vertex to a downstream vertex.}
\usage{
plot_path(g, src, trg, main = NULL)
}
\arguments{
\item{g}{igraph object}

\item{src}{an upstream vertex in g}

\item{trg}{a downstream vertex in g}

\item{main}{character, the desired title of the graph. Default is NULL}
}
\description{
Plot path from an upstream vertex to a downstream vertex.
}
\examples{
set.seed(20)
g <- layer_DAGs(6, 3)
plot_path(g, 4, 11)
}

