% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/propagation.R
\name{updateVertices}
\alias{updateEdges}
\alias{updateVertices}
\title{Vertex/Edge Propagation in igraph}
\usage{
updateVertices(g, getDeterminers, callback)

updateEdges(g, getDeterminers, callback)
}
\arguments{
\item{g}{A graph object.}

\item{getDeterminers}{The function that gives the vertices/edges that must be updated for the callback to be
executed}

\item{callback}{The function performed on each vertex/edge}
}
\value{
A igraph object with updated vertex/edge states. A warning message is returned if not all vertices
could be updated.
}
\description{
Foundation for any graph propagation algorithm, where the states of vertexes or edges propagate through the graph,
changing the states of others.  Traverse the graph, land on a vertex/edge, perform a callback function on the
vertex/edge.  That callback can depend on the states of other 'determiner' vertices/edges.  When the callback is
completed successfully, the 'updated' attribute of the object is changed to \code{TRUE}. If those determiners do
not have the status \code{updated == TRUE}, then the function recursively performs the callback on the determiners until all
their states are updated, before finalling performing the callback on the original vertex/edge.
}
\details{
\code{updateVertices} performs propagation on vertices. \code{updateEdges} performs propagation on edges.
}
\section{Functions}{
\itemize{
\item \code{updateEdges}: 
}}

